+++ src/worker/03-write-db.js
@@ -1,0 +1,45 @@
+const chalk = require('chalk')
+const openDB = require('../lib/open-db')
+const fns = require('../lib/fns')
+const mongoConfig = {
+  ordered: false
+}
+
+//report how many pages we wrote this time
+const writeMsg = function(pages, count, start, workerNum) {
+  let msg = chalk.yellow(` #${workerNum}  `)
+  count = fns.niceNumber(count)
+  msg += chalk.green(`+${count} `) + 'pages'
+  msg += chalk.grey('  -  ' + fns.timeSince('  ' + start))
+  msg += chalk.blue(` - `)
+  msg += chalk.magenta(`"${pages[0].title}"`)
+  console.log(msg)
+}
+
+const writeDb = async (options, pages, workerNum) => {
+  const start = Date.now()
+  let obj = await openDB(options)
+
+  let result = await obj.col.insertMany(pages, mongoConfig).catch(async (err) => {
+    if (err.code === 11000) {
+      let errCount = err.result.getWriteErrorCount()
+      errCount = fns.niceNumber(errCount)
+      console.log(chalk.red(`-- ${errCount}  duplicate pages --`))
+    } else {
+      console.log(chalk.red(`====error!===`))
+      console.log(err)
+    }
+    err = err.result.toJSON()
+    const count = err.nInserted
+    writeMsg(pages, count, start, workerNum)
+    await obj.client.close()
+  })
+  //no errors thrown, all good
+  if (result) {
+    const count = result.insertedCount
+    writeMsg(pages, count, start, workerNum)
+    await obj.client.close()
+  }
+}
+
+module.exports = writeDb